% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_mm}
\alias{atac_mm}
\title{\code{atac_mm} differential binding results}
\format{
A DBA object (11 MGUS samples and 11 MM tumor samples, 65114 sites)
}
\source{
\href{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA1174220}{PRJNA1174220} (\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE279761}{GSE279761})
}
\usage{
atac_mm
}
\description{
\code{DiffBind} differential analyses of ATAC-seq data from 11 Monoclonal Gammopathy of Undetermined Significance (MGUS) and 11 Multiple Myeloma (MM) human samples obtained from T. Bruno, M.C. Cappelletto, C. Cortile et al. (Blood, 2025).
}
\references{
T. Bruno, M.C. Cappelletto, C. Cortile et al. (Blood, 2025) \doi{https://doi.org/10.1182/blood.2025028441}
}
\keyword{datasets}
