% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atac_mm_SraRunTable}
\alias{atac_mm_SraRunTable}
\title{\code{atac_mm_SraRunTable} metadata table}
\format{
data.frame (66 rows x 35 columns)
}
\source{
\href{https://www.ncbi.nlm.nih.gov/bioproject/PRJNA1174220}{PRJNA1174220} (\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE279761}{GSE279761})
}
\usage{
atac_mm_SraRunTable
}
\description{
metadata table for the download of the fastq ATAC-seq data from T. Bruno, M.C. Cappelletto, C. Cortile et al. (Blood, 2025).
}
\references{
T. Bruno, M.C. Cappelletto, C. Cortile et al. (Blood, 2025) \doi{https://doi.org/10.1182/blood.2025028441}
}
\keyword{datasets}
