% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endometrium}
\alias{endometrium}
\title{\code{endometrium} differential binding results}
\format{
A DBA object (4 normal samples and 4 tumor samples)
}
\source{
Gregoricchio et al. (Genome Biology, 2025) \url{https://doi.org/10.1186/s13059-025-03596-5}
}
\usage{
endometrium
}
\description{
DiffBind differential analyses of ERa ChIP-seq data performed in normal and tumor endometrium from human patients, Gregoricchio et al. (Genome Biology, 2025).
}
\keyword{datasets}
