% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.stats.R
\name{get.stats}
\alias{get.stats}
\title{get.stats}
\usage{
get.stats(
  dba.object,
  contrast,
  design,
  diff.method = "DESeq2",
  p.adjust.method = "BH"
)
}
\arguments{
\item{dba.object}{An object of class \code{DBA} obtained from DiffBind.}

\item{contrast}{Vector of 3-elements indicating (in order): metadata_column, group_1, group_2.}

\item{design}{String indicating the design to use in the differential analyses.}

\item{diff.method}{String indicating which tool to use to perform differential analyses. One among: 'DESeq2' and 'edgeR'. Default: \code{"DESeq2"}.}

\item{p.adjust.method}{String indicating the method to use to correct the p-values. One among: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: \code{BH}.}
}
\value{
An object of class \code{power4omics.stats}.
}
\description{
Extract statistics and degrees of freedom from a DiffBind object (class \code{DBA}).
}
