% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,power4omics.power-method}
\alias{show,power4omics.power-method}
\title{power4omics.power show-method}
\usage{
\S4method{show}{power4omics.power}(object)
}
\description{
power4omics.power show-method
}
