% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,power4peaks.power-method}
\alias{show,power4peaks.power-method}
\title{power4peaks.power show-method}
\usage{
\S4method{show}{power4peaks.power}(object)
}
\arguments{
\item{object}{Object of class \code{power4peaks.power}.}
}
\description{
power4peaks.power show-method
}
