% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.and.methods.R
\name{show,power4peaks.stats-method}
\alias{show,power4peaks.stats-method}
\title{power4peaks.stats show-method}
\usage{
\S4method{show}{power4peaks.stats}(object)
}
\description{
power4peaks.stats show-method
}
