% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.power.R
\name{compute.power}
\alias{compute.power}
\title{compute.power}
\usage{
\method{compute}{power}(
  power4omics.stats,
  sample.size.range = c(0, 50),
  power.threshold = 0.8,
  distribution.type = NULL,
  df = NULL
)
}
\arguments{
\item{power4omics.stats}{An object of class \code{power4omics.stats}.}

\item{sample.size.range}{Numeric vector of length 2, indicating minimum and maximum of the range of number of samples in which test the power. Default: \code{c(0,50)}.}

\item{power.threshold}{number between 0 and 1 indicating the threshold to plot for the power. Default: \code{0.8} (80 percent power).}

\item{distribution.type}{Integer indicating the number of degrees of freedom. Default: \code{NULL}, automatically defined from the power4omics.stats object.}
}
\value{
An object of class \code{power4omics.power}.
}
\description{
Compute the power and effect size for the object obtained from \link{get.stats}.
}
