% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stat.distribution.R
\name{plot.stat.distribution}
\alias{plot.stat.distribution}
\title{plot.stat.distribution}
\usage{
\method{plot}{stat.distribution}(
  power4peaks.stats,
  obs.stat.color = "black",
  norm.fit.color = "firebrick",
  chi.fit.color = "navy",
  f.fit.volor = "forestgreen"
)
}
\arguments{
\item{power4peaks.stats}{An object of class \code{power4peaks.stats}.}

\item{obs.stat.color}{Any R-supported color string indicating the color to use for the observed statistics distribution. Default: \code{"black"}.}

\item{norm.fit.color}{Any R-supported color string indicating the color to use for the normal distribution. Default: \code{"firebrick"}.}

\item{chi.fit.color}{Any R-supported color string indicating the color to use for the chi-squared distribution. Default: \code{"navy"}.}

\item{f.fit.color}{Any R-supported color string indicating the color to use for the F distribution. Default: \code{"forestgreen"}.}
}
\value{
A ggplot object.
}
\description{
Plot the distribution of the statistics obtained from \link{get.stats}. Normal and chi-squared distributions are overlayed.
}
