% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stat.distribution.R
\name{plot.stat.distribution}
\alias{plot.stat.distribution}
\title{plot.stat.distribution}
\usage{
\method{plot}{stat.distribution}(
  power4peaks.stats,
  obs.stat.color = "black",
  norm.fit.color = "firebrick",
  chi.fit.color = "navy",
  t.fit.color = "violet",
  f.fit.color = "forestgreen",
  df1 = NULL,
  df2 = NULL,
  adapt.y.axis = TRUE
)
}
\arguments{
\item{power4peaks.stats}{An object of class \code{power4peaks.stats}.}

\item{obs.stat.color}{Any R-supported color string indicating the color to use for the observed statistics distribution. Default: \code{"black"}.}

\item{norm.fit.color}{Any R-supported color string indicating the color to use for the normal distribution. Default: \code{"firebrick"}.}

\item{chi.fit.color}{Any R-supported color string indicating the color to use for the chi-squared distribution. Default: \code{"navy"}.}

\item{t.fit.color}{Any R-supported color string indicating the color to use for the t distribution. Default: \code{"violet"}.}

\item{f.fit.color}{Any R-supported color string indicating the color to use for the F distribution. Default: \code{"forestgreen"}.}

\item{df1}{Numeric value forcing the value for the degrees-of-freedom 1. Default: \code{NULL} (the value is obtained from the \code{power4peaks.stats} object.}

\item{df2}{Numeric value forcing the value for the degrees-of-freedom 2. Default: \code{NULL} (the value is obtained from the \code{power4peaks.stats} object.}

\item{adapt.y.axis}{Logical value indicating whether the Y-axis should be scaled to fit the observed distribution/ Default: \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
Plot the distribution of the statistics obtained from \link{get.stats}. Normal and chi-squared distributions are overlayed.
}
